backend/services/extractor.py
from pypdf import PdfReader

def extract_pdf_text(file):
    reader = PdfReader(file)
    text = ""
    for page in reader.pages:
        text += page.extract_text() or ""
    return text

import requests
from bs4 import BeautifulSoup

def extract_url_text(url):
    r = requests.get(url)
    soup = BeautifulSoup(r.text, "html.parser")
    return soup.get_text()